code structure changed + migrations changes done

+++ database/migrations/2020_08_24_183944_create_permissions_table.php
@@ -15,7 +15,7 @@ class CreatePermissionsTable extends Migration
     {
         Schema::create('permissions', function (Blueprint $table) {
             $table->id();
-            $table->integer('role_id');
+            $table->bigInteger('role_id');
             $table->string('slug');
             $table->timestamps();
         });
